requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM
google-generativeai==0.3.2
ollama==0.1.6
chromadb==0.4.22
numpy<2.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Email Processing
email-validator==2.1.0
python-multipart==0.0.6
extract-msg==0.45.0

# File Processing
python-magic==0.4.27
pypdf==3.17.4
python-docx==1.1.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Google Workspace
google-api-python-client==2.111.0
google-auth==2.27.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
